corrected mistakes

+++ src/main/java/com/example/doctors_guide/controllers/DiseaseNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.doctors_guide.controllers;
 
 public class DiseaseNotFoundException extends RuntimeException {
 
-    public DiseaseNotFoundException(Long id) {
+    public DiseaseNotFoundException(Integer id) {
         super("Could not find disease with id: " + id);
     }
 }
